requirements.txt
alabaster==0.7.16
asgiref==3.11.1
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==6.0.2
django-environ==0.12.0
django-unfold==0.79.0
djangorestframework==3.16.1
docutils==0.20.1
gunicorn==25.0.3
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.6.0
packaging==24.0
pexpect==4.9.0
ptyprocess==0.7.0
Pygments==2.17.2
requests==2.31.0
setuptools==69.2.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.5
urllib3==2.2.1
wheel==0.43.0
